% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\docType{data}
\name{catch}
\alias{catch}
\title{Catch data}
\format{A data frame with 6 rows and 7 variables:
\itemize{
\item fish species
\item number of fish of that species catch at location A
\item number of fish of that species catch at location B
\item number of fish of that species catch at location C
\item number of fish of that species catch at location D
\item number of fish of that species catch at location E
\item number of fish of that species catch at location F
}}
\source{
Ignacia Rivera
}
\usage{
catch
}
\description{
Data of catch for different locations in Sanata Barbara for different species
}
\keyword{datasets}
