% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_function.R
\name{fish_function}
\alias{fish_function}
\title{Fish function}
\usage{
fish_function(prices = NULL, catch = NULL, plot = FALSE)
}
\arguments{
\item{prices}{a table that has prices for different fish}

\item{catch}{a table that has the number caught for each fish species for each location. Each location is in a different column, and each fish is in a different row.}

\item{plot}{a logical (default = FALSE) indicating if a graph of revenue by location and total revenue must be produced}
}
\value{
a list containing the following elements: 1) a tibble (frequent_fish) containing the most frequently caught fish in each location 2) a tibble (site_revenue) containing the total revenue for each location 3) an integer (total_catches) with the total fisheries sum, and 4) a plot (plot) of revenue by location
}
\description{

}
\author{
Lina Barbosa, Ignacia Rivera & Rucha Thakar
}
