% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_price_analysis.R
\name{fish_price_analysis}
\alias{fish_price_analysis}
\title{Fish price analysis}
\usage{
fish_price_analysis(prices = NULL)
}
\arguments{
\item{prices}{a table that has prices for different fish}
}
\value{
a list containing the following elements: 1) the value of the average price across different fish species (average_price) 2) a tibble indicating the least valued species and its price (least_valued) 3) a tibble indicating the most valued species and its price (most_valued)
}
\description{

}
\author{
Lina Barbosa, Ignacia Rivera & Rucha Thakar
}
