% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\docType{data}
\name{price}
\alias{price}
\title{Price data}
\format{A data frame with 6 rows and 7 variables:
\itemize{
\item fish species
\item price for the specie
}}
\source{
Ignacia Rivera
}
\usage{
price
}
\description{
Data of price for different fish species in Sanata Barbara dock
}
\keyword{datasets}
